Move chapter 1 listings to standard file layout.
