--- src/main/kotlin/chapter1/Listing_1_2.kt
@@ -1,21 +1,0 @@
-package chapter1
-
-object Listing_1_2 {
-    data class Coffee(val price: Float = 2.95F)
-
-    class CreditCard
-
-    class Payments {
-        fun charge(cc: CreditCard, price: Float): Unit = TODO()
-    }
-
-    //tag::init[]
-    class Cafe {
-        fun buyCoffee(cc: CreditCard, p: Payments): Coffee {
-            val cup = Coffee()
-            p.charge(cc, cup.price)
-            return cup
-        }
-    }
-    //end::init[]
-}
