--- src/main/kotlin/chapter1/Listing_1_3.kt
@@ -1,19 +1,0 @@
-package chapter1
-
-object Listing_1_3 {
-
-    class CreditCard
-
-    data class Coffee(val price: Float = 2.50F)
-
-    data class Charge(val cc: CreditCard, val amount: Float)
-
-    //tag::init[]
-    class Cafe {
-        fun buyCoffee(cc: CreditCard): Pair<Coffee, Charge> {
-            val cup = Coffee()
-            return Pair(cup, Charge(cc, cup.price))
-        }
-    }
-    //end::init[]
-}
